Fix: Improve logic to handle mood frequencies

+++ app/src/main/java/com/raulastete/aura/core/features/statistics/MoodStatistics.kt
@@ -11,5 +11,5 @@ data class HeatmapDay(
 data class MoodFrequency(
     val mood: Mood,
     val count: Int,
-    val percentage: Float
+    val percentage: Double
 )
